crypto/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
backports.pbkdf2==0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crypto==1.4.1
idna==3.7
Naked==0.1.32
pbkdf2==1.3
pyaes==1.6.1
pycparser==2.22
pycryptodome==3.20.0
pyDHE==1.0.0
pyscrypt==1.6.2
PyYAML==6.0.1
requests==2.32.2
scrypt==0.8.24
shellescape==3.8.1
urllib3==2.2.1
